charmodel: validation and text-cross-entropy share code
Both now use doubles for accumulating cross-entropy, so that much
larger texts can be accurately measured (the cumulative sum can make the
difference between the sum and each character's entropy large).

+++ charmodel.h
@@ -199,7 +199,7 @@ RnnCharAlphabet *rnn_char_new_alphabet_from_net(RecurNN *net);
 void rnn_char_alphabet_set_flags(RnnCharAlphabet *a,
     bool case_insensitive, bool utf8, bool collapse_space);
 
-float rnn_char_cross_entropy(RecurNN *net, RnnCharAlphabet *alphabet,
+double rnn_char_cross_entropy(RecurNN *net, RnnCharAlphabet *alphabet,
     const u8 *text, const int len, const int ignore_first,
     const u8 *prefix_text, const int prefix_len);
